Modified the CRUD functions to fetch and retrieve data from the view

+++ AssetTracking/AssetTracking/Controllers/OfficeBooksController.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using AssetTracking.Helpers;
@@ -8,53 +9,95 @@ using Microsoft.Graph;
 
 namespace AssetTracking.Controllers
 {
-    [Route("[controller]")]
+    //[Route("[controller]")]
     public class OfficeBooksController : Controller
     {
         private readonly IGraphSdkHelper _graphSdkHelper;
         private IOfficeBookRepository _officeBookRepository;
         private GraphServiceClient _graphClient;
 
-        public  OfficeBooksController( IGraphSdkHelper graphSdkHelper, IOfficeBookRepository officeBookRepository) 
-        { 
+        public OfficeBooksController(IGraphSdkHelper graphSdkHelper, IOfficeBookRepository officeBookRepository)
+        {
             _graphSdkHelper = graphSdkHelper;
             _officeBookRepository = officeBookRepository;
         }
+
+        //[HttpGet]
+        public ActionResult OfficeBooks()
+        {
             
-        [HttpGet]
-        public async Task<IActionResult> OfficeBooks()
+            return View();
+        }
+
+        public async Task<JsonResult> OfficeBooksGet()
         {
             if (User.Identity.IsAuthenticated)
             {
                 _graphClient = _graphSdkHelper.GetAuthenticatedClient((ClaimsIdentity)User.Identity);
                 List<OfficeBook> officeBook = await _officeBookRepository.GetBooks(_graphClient);
-                ViewBag.List = officeBook;
-                return View(officeBook);
+
+                return Json(new { data = officeBook });
             }
-            return View();
+            else
+            {
+                return Json(new { IsSuccess = false });
+            }
+        }
+
+        public async Task<ActionResult> OfficeBooksGetbyId(string Id)
+        {
+            if (User.Identity.IsAuthenticated)
+            {
+                _graphClient = _graphSdkHelper.GetAuthenticatedClient((ClaimsIdentity)User.Identity);
+                List<OfficeBook> officeBookList = await _officeBookRepository.GetBooks(_graphClient);
+                OfficeBook officeBook = officeBookList.Where(d => d.ItemId == Id).FirstOrDefault();
+                return Json(officeBook);
+            }
+            return Json(null);
         }
 
         [HttpPost]
-        [AutoValidateAntiforgeryToken]
-        public async Task<IActionResult> AddBook(OfficeBook officeBook)
+        public async Task<JsonResult> AddBook(OfficeBook officeBook)
         {
-            _graphClient = _graphSdkHelper.GetAuthenticatedClient((ClaimsIdentity)User.Identity);
-            bool result = await _officeBookRepository.AddBook(officeBook,_graphClient);
-            return RedirectToAction("~/Views/OfficeBooks/OfficeBooks.cshtml");
+            if (User.Identity.IsAuthenticated)
+            {
+                _graphClient = _graphSdkHelper.GetAuthenticatedClient((ClaimsIdentity)User.Identity);
+                bool result = await _officeBookRepository.AddBook(officeBook, _graphClient);
+
+                return Json(new { IsSuccess = result });
+            }
+            else
+            {
+                return Json(new { IsSuccess = false });
+            }
         }
 
-        [HttpPut]
-        public async Task<bool> UpdateBook(OfficeBook officeBook)
+        public async Task<JsonResult> UpdateBook(OfficeBook officeBook)
         {
-            bool result = await _officeBookRepository.UpdateBook(officeBook, _graphClient);
-            return result;
+            if (User.Identity.IsAuthenticated)
+            {
+                _graphClient = _graphSdkHelper.GetAuthenticatedClient((ClaimsIdentity)User.Identity);
+                bool result = await _officeBookRepository.UpdateBook(officeBook, _graphClient);
+                return Json(new { IsSuccess = result });
+            }
+            else
+            {
+                return Json(new { IsSuccess = false });
+            }
         }
-        [Route("/DeleteBook")]
-        [HttpPost]
-        public async Task<bool> DeleteBook(OfficeBook officeBook)
+
+        public async Task<JsonResult> DeleteBook(OfficeBook officeBook)
         {
-            bool result = await _officeBookRepository.DeleteBook(officeBook, _graphClient);
-            return result;
+            if (User.Identity.IsAuthenticated)
+            {
+                _graphClient = _graphSdkHelper.GetAuthenticatedClient((ClaimsIdentity)User.Identity);
+                bool result = await _officeBookRepository.DeleteBook(officeBook, _graphClient);
+                return Json(new { IsSuccess = result });
+            }
+            else
+            {
+                return Json(new { IsSuccess = false });
+            }
         }
     }
 }
